+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Garden Features
 + Added **Copper Price** - Show copper to coin prices inside the Sky Mart inventory.
 + Added **Visitor Display** - Show all items needed for the visitors.
-+ Added **Highlight Ready** - Highlight the visitor when the required items are in the inventory.
++ Added **Visitor Highlight** - Highlight visitor when the required items are in the inventory or the visitor is new and needs to checked what items it needs.
 + Added **Show Price** - Show the bazaar price of the items required for the visitors.
 + Added **Crop Milestone** Number - Show the number of the crop milestone in the inventory.
 + Added **Crop Upgrades** Number - Show the number of upgrades in the crop upgrades inventory.